Make the relevance vector averaging cumulative
The relevance vector (i.e. the average vector for a result marked
relevant by the user) is now computed cumulatively across all of the
queries.  This assumes the user has a consistent relevance criteria in
mind, which seems reasonable, and allows us to consider knowledge from
previous queries to improve the current one.